flask-server/requirements.txt
Flask
flask-cors
python-dotenv
google-generativeai
gunicorn

